+++ CppApplication_1/vars.h
@@ -24,7 +24,7 @@ extern int idx_prc;
 extern int nodos;
 extern int idx_vec;
 extern int idx_vec2;
-extern int* arrayVectores [32];
+extern double* arrayVectores [32];
 extern char* arrayVectoresAlfa [32];
 extern int depurar;
 //#endif	// VARS_H
